requirements.txt
aniso8601==7.0.0
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
dominate==2.3.5
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pycparser==2.19
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
